One of casts is removed
git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@9312 9d14d920-75d5-0310-8bd7-a52572981aee

+++ lib/tapi_rpcsock/tapi_rpcsock.c
@@ -3579,7 +3579,7 @@ rpc_ioctl(rcf_rpc_server *handle,
             {
                 in.req.req_val[0].type = IOCTL_IFREQ;
                 in.req.req_val[0].ioctl_request_u.req_ifreq.rpc_ifr_flags = 
-                    if_fl_h2rpc((uint32_t)(unsigned short int)((struct ifreq *)arg)->ifr_flags);
+                    if_fl_h2rpc((uint32_t)((struct ifreq *)arg)->ifr_flags);
                 memcpy(in.req.req_val[0].ioctl_request_u.req_ifreq.
                            rpc_ifr_name,
                        ((struct ifreq *)arg)->ifr_name,
